+++ busy_beaver/apps/events/workflow.py
@@ -3,7 +3,7 @@ import time
 import click
 
 from .blueprint import events_bp
-from busy_beaver.apps.upcoming_events.cards import UpcomingEventList
+from busy_beaver.apps.events.cards import UpcomingEventList
 from busy_beaver.clients import SlackClient
 from busy_beaver.common.wrappers.meetup import EventDetails
 from busy_beaver.models import Event, SlackInstallation
