+++ busy_beaver/apps/slack_integration/slash_command.py
@@ -3,14 +3,14 @@ from typing import List, NamedTuple
 
 from .decorators import limit_to
 from .toolbox import make_slack_response
+from busy_beaver.apps.events.workflow import (
+    generate_next_event_message,
+    generate_upcoming_events_message,
+)
 from busy_beaver.apps.github_integration.oauth.workflow import (
     connect_github_to_slack,
     disconnect_github_from_slack,
     relink_github_to_slack,
-)
-from busy_beaver.apps.upcoming_events.workflow import (
-    generate_next_event_message,
-    generate_upcoming_events_message,
 )
 from busy_beaver.config import FULL_INSTALLATION_WORKSPACE_IDS, MEETUP_GROUP_NAME
 from busy_beaver.toolbox import EventEmitter
