+++ busy_beaver/models.py
@@ -1,5 +1,6 @@
 from busy_beaver.common.models import BaseModel, KeyValueStore, Task  # noqa isort:skip
 
+from busy_beaver.apps.events.models import Event  # noqa
 from busy_beaver.apps.github_integration.models import (  # noqa
     GitHubSummaryConfiguration,
     GitHubSummaryUser,
@@ -8,5 +9,4 @@ from busy_beaver.apps.slack_integration.models import (  # noqa
     SlackAppHomeOpened,
     SlackInstallation,
 )
-from busy_beaver.apps.upcoming_events.models import Event  # noqa
 from busy_beaver.apps.youtube_integration.models import YouTubeVideo  # noqa
